Grab footer copyright from correct location
The site.yaml contains 2 variables with the copyright text: 1 under author and 1 under footer.

It makes no sense to have this variable under author, not to define it 2x. Therefore, for clarity and standardization and easier maintenance, site.yaml has been updated as well, to provide this single source of truth for the copyright text.